remove old GUI and rename new GUI files
